feat: add dynamic naming, UI persistency, date in path

+++ src/pipelines/alertwildfire.py
@@ -1,4 +1,12 @@
+import os
+import sys
+
+libraries_path = os.path.abspath(os.path.join(os.path.dirname(__file__), ".."))
+sys.path.append(libraries_path)
+
+
 import json
+import logging  # Import the logging module
 import os
 import shutil
 from datetime import datetime
@@ -6,7 +14,7 @@ from datetime import datetime
 import luigi
 import requests
 
-from dl_images import download_and_process_camera, get_camera_local_time
+from libraries.dl_images import download_and_process_camera, get_camera_local_time
 
 # Constants
 CAMERAS_URL = (
@@ -20,7 +28,14 @@ HEADERS = {
     "Referer": "https://www.alertwildfire.org/",
     "Host": "s3-us-west-2.amazonaws.com",
 }
-OUTPUT_BASE_PATH = "./"
+OUTPUT_BASE_PATH = "./alertwildfire_data/"
+
+# Setup logging
+logging.basicConfig(
+    level=logging.INFO,  # You can change to DEBUG if you need more detailed logs
+    format="%(asctime)s [%(levelname)s] %(message)s",
+    handlers=[logging.StreamHandler()],
+)
 
 
 class FetchCamerasData(luigi.Task):
@@ -33,6 +48,7 @@ class FetchCamerasData(luigi.Task):
         return luigi.LocalTarget(os.path.join(OUTPUT_BASE_PATH, "cameras_data.json"))
 
     def run(self):
+        logging.info("Fetching cameras data from API...")
         response = requests.get(CAMERAS_URL, headers=HEADERS)
         response.raise_for_status()  # Ensure we raise an error for bad responses
         cameras_data = response.json()
@@ -40,6 +56,7 @@ class FetchCamerasData(luigi.Task):
         # Write camera data to a JSON file
         with self.output().open("w") as f:
             json.dump(cameras_data, f)
+        logging.info(f"Camera data saved to {self.output().path}")
 
 
 class DownloadImagesForCamera(luigi.Task):
@@ -50,33 +67,45 @@ class DownloadImagesForCamera(luigi.Task):
 
     camera_feature = luigi.DictParameter()
 
-    def output(self):
-        state = self.camera_feature["properties"].get("state")
-        source = self.camera_feature["properties"].get("id").lower()
-        local_time = get_camera_local_time(state)
-        output_path = os.path.join(
-            OUTPUT_BASE_PATH, "temp", local_time.strftime("%Y_%m_%d")
+    def __init__(self, *args, **kwargs):
+        super(DownloadImagesForCamera, self).__init__(*args, **kwargs)
+        self.state = self.camera_feature["properties"].get("state")
+        self.source = self.camera_feature["properties"].get("id").lower()
+        self.local_time = get_camera_local_time(self.state)
+        self.output_path = os.path.join(
+            OUTPUT_BASE_PATH, "temp", self.local_time.strftime("%Y_%m_%d-%H")
         )
-        camera_path = os.path.join(output_path, source)
-        marker_file = os.path.join(camera_path, "download_complete.txt")
+        self.camera_path = os.path.join(self.output_path, self.source)
+
+    @property
+    def task_family(self):
+        # Dynamically define task name based on camera ID
+        camera_id = self.camera_feature["properties"].get("id").lower()
+        return f"Dl-cam-{camera_id}"
+
+    def output(self):
+        marker_file = os.path.join(self.camera_path, "download_complete.txt")
         return luigi.LocalTarget(marker_file)
 
     def run(self):
-        state = self.camera_feature["properties"].get("state")
-        source = self.camera_feature["properties"].get("id").lower()
-        local_time = get_camera_local_time(state)
-        output_path = os.path.join(
-            OUTPUT_BASE_PATH, "temp", local_time.strftime("%Y_%m_%d")
+        logging.info(
+            f"Downloading images for camera {self.source} (state: {self.state})..."
         )
-        camera_path = os.path.join(output_path, source)
 
         # Download and process images for the camera
-        download_and_process_camera(self.camera_feature["properties"])
+        download_and_process_camera(self.camera_feature["properties"], self.output_path)
 
         # Create the marker file to indicate completion
-        os.makedirs(camera_path, exist_ok=True)
-        with self.output().open("w") as f:
-            f.write("Download complete.\n")
+        if os.path.exists(self.camera_path) and os.listdir(self.camera_path):
+            with self.output().open("w") as f:
+                f.write("Download complete.\n")
+            logging.info(
+                f"Download complete for camera {self.source}. Marker created at {self.output().path}"
+            )
+        else:
+            raise RuntimeError(
+                f"Le dossier {self.camera_path} est vide ou inexistant après le téléchargement."
+            )
 
 
 class ProcessImagesForCamera(luigi.Task):
@@ -87,6 +116,12 @@ class ProcessImagesForCamera(luigi.Task):
 
     camera_feature = luigi.DictParameter()
 
+    @property
+    def task_family(self):
+        # Dynamically define task name based on camera ID
+        camera_id = self.camera_feature["properties"].get("id").lower()
+        return f"Process-cam-{camera_id}"
+
     def requires(self):
         return DownloadImagesForCamera(camera_feature=self.camera_feature)
 
@@ -95,7 +130,7 @@ class ProcessImagesForCamera(luigi.Task):
         source = self.camera_feature["properties"].get("id").lower()
         local_time = get_camera_local_time(state)
         processed_path = os.path.join(
-            OUTPUT_BASE_PATH, "processed", local_time.strftime("%Y_%m_%d"), source
+            OUTPUT_BASE_PATH, "processed", local_time.strftime("%Y_%m_%d-%H"), source
         )
         marker_file = os.path.join(processed_path, "process_complete.txt")
         return luigi.LocalTarget(marker_file)
@@ -105,11 +140,13 @@ class ProcessImagesForCamera(luigi.Task):
         source = self.camera_feature["properties"].get("id").lower()
         local_time = get_camera_local_time(state)
         download_path = os.path.join(
-            OUTPUT_BASE_PATH, "temp", local_time.strftime("%Y_%m_%d"), source
+            OUTPUT_BASE_PATH, "temp", local_time.strftime("%Y_%m_%d-%H"), source
         )
         processed_path = os.path.join(
-            OUTPUT_BASE_PATH, "processed", local_time.strftime("%Y_%m_%d"), source
+            OUTPUT_BASE_PATH, "processed", local_time.strftime("%Y_%m_%d-%H"), source
         )
+
+        logging.info(f"Processing images for camera {source}...")
 
         # Ensure the processed directory exists
         os.makedirs(processed_path, exist_ok=True)
@@ -130,9 +167,9 @@ class ProcessImagesForCamera(luigi.Task):
             if self.apply_algorithm(image_path):
                 # Move processed images to the 'processed' directory
                 shutil.move(image_path, os.path.join(processed_path, image_file))
+                logging.info(f"Processed and moved image: {image_file}")
             else:
-                # In case of failure, you can skip or log the error.
-                print(f"Error processing {image_file}")
+                logging.error(f"Error processing {image_file}")
 
         # Optionally, remove the original folder once processing is complete
         if os.path.exists(download_path) and not os.listdir(download_path):
@@ -141,6 +178,9 @@ class ProcessImagesForCamera(luigi.Task):
         # Create the marker file to indicate completion
         with self.output().open("w") as f:
             f.write("Process complete.\n")
+        logging.info(
+            f"Processing complete for camera {source}. Marker created at {self.output().path}"
+        )
 
     def apply_algorithm(self, image_path):
         """
@@ -157,18 +197,29 @@ class RunCameraWorkflow(luigi.WrapperTask):
     Wrapper task to run the entire workflow for all cameras.
     """
 
+    @property
+    def task_family(self):
+        # Dynamically define task name based on camera ID
+        return "Cam-Workflow-" + str(datetime.now().strftime("%Y_%m_%d-%H"))
+
     def requires(self):
         # First, fetch camera data
-        return FetchCamerasData()
+        yield FetchCamerasData()
+        # Define a generator to yield all ProcessImagesForCamera tasks
+        cameras_file_path = os.path.join(OUTPUT_BASE_PATH, "cameras_data.json")
+        if os.path.exists(cameras_file_path):
+            with open(cameras_file_path) as f:
+                cameras = json.load(f)
+
+            # Yield tasks for each camera
+            for camera_feature in cameras["features"][:2]:
+                logging.info(
+                    f"Scheduling processing for camera {camera_feature['properties'].get('id')}"
+                )
+                yield ProcessImagesForCamera(camera_feature=camera_feature)
 
     def run(self):
-        # Define a generator to yield all ProcessImagesForCamera tasks
-        with self.input().open("r") as f:
-            cameras = json.load(f)
-
-        # Yield tasks for each camera
-        for camera_feature in cameras["features"][:10]:
-            yield ProcessImagesForCamera(camera_feature=camera_feature)
+        print("habile")
 
 
 if __name__ == "__main__":
